LPD-57956 Merge IndexFactory/CompanyIndexFactory into a pojo
